refactoring queries
