+++ src/components/pages/Stats/BarChart.vue
@@ -31,7 +31,7 @@ import { Component, Mixins, Prop } from 'vue-property-decorator';
 import ChartSpecMixin from '@/components/mixins/ChartSpecMixin';
 import { Components } from '@/consts';
 import { SECONDS_IN_TYPE, NETWORK_STATS_FILTERS } from '@/consts/snapshots';
-import { fetchData } from '@/indexer/queries/networkVolume';
+import { fetchData } from '@/indexer/queries/network/volume';
 import { lazyComponent } from '@/router';
 import { getter } from '@/store/decorators';
 import type { SnapshotFilter } from '@/types/filters';
