+++ src/indexer/queries/network/stats.ts
